+++ simulation/Makefile
@@ -1,20 +1,25 @@
-.PHONY: run push-to-ecr build-client
+.PHONY: run build arkd push-to-ecr build-client run-web vet
 
 #### Local simulation ####
-## run-single: Run bundled simulation where client and server are in the same process
-run-single:
+## run: Run simulation locally Regtest
+run:
 	@echo "Running singe process simulation..."
-	go run ./local/single-process/main.go
+	go run ./local/main.go $(ARGS)
 
-## run-multi: Run simulation with separate client and server processes
-run-multi: build-local-client
-	@echo "Running multi-process simulation..."
-	go run ./local/process-per-client/main.go
+## build:
+build:
+	@echo "Building simulation..."
+	go build -o ./build/client ./local/client/main.go
 
-## build-local-client
-build-local-client:
-	@echo "Building local client..."
-	go build -o ./build/client ./local/process-per-client/client/main.go
+## setup-ark:
+setup-ark: arkd
+	@echo "Setup arkd wallet..."
+	./script/setup_arkd_wallet.sh
+
+## arkd:
+arkd:
+	@echo "Starting arkd..."
+	cd ../server && make run &
 
 #### Remote simulation ####
 
@@ -34,25 +39,12 @@ build-client:
 	@echo "Building Docker image without pushing to ECR..."
 	./script/build_and_push.sh -a $(AWS_ACCOUNT_ID) -r $(AWS_REGION) -e $(ECR_REPOSITORY_NAME) -i $(IMAGE_NAME) -d $(DOCKERFILE_PATH) -c $(BUILD_CONTEXT) --no-push
 
-## run-remote: Run simulation on remote server
-run-remote:
-	@echo "Running simulation on remote server..."
-	@if [ -z "$(SUBNET_IDS)" ]; then \
-		echo "Error: SUBNET_IDS is not set. You can find this in CloudFormation outputs."; \
-		exit 1; \
-	fi
-	@if [ -z "$(SECURITY_GROUP_IDS)" ]; then \
-		echo "Error: SECURITY_GROUP_IDS is not set. You can find this in CloudFormation outputs."; \
-		exit 1; \
-	fi
-	go run ./remote/orchestrator/main.go $(ARGS)
-
-vet:
-	@echo "Running go vet..."
-	go vet ./...
-
-#### Web Application ####
 ## run-web: Run web interface for simulation
 run-web:
 	@echo "Starting web interface..."
 	go run ./web/main.go $(ARGS)
+
+#### Code Analysis ####
+vet:
+	@echo "Running go vet..."
+	go vet ./...
