+++ simulation/config/schema.yaml
@@ -1,0 +1,61 @@
+$schema: "http://json-schema.org/draft-07/schema#"
+type: object
+required:
+  - version
+  - clients
+  - rounds
+properties:
+  version:
+    type: string
+  description:
+    type: string
+  clients:
+    type: array
+    items:
+      type: object
+      required:
+        - id
+      properties:
+        id:
+          type: string
+        name:
+          type: string
+        initial_funding:
+          type: number
+          minimum: 0
+  rounds:
+    type: array
+    items:
+      type: object
+      required:
+        - number
+        - actions
+      properties:
+        number:
+          type: integer
+          minimum: 1
+        actions:
+          type: object
+          patternProperties:
+            "^client_\\d+$":
+              type: array
+              items:
+                type: object
+                required:
+                  - type
+                properties:
+                  type:
+                    type: string
+                    enum:
+                      - "Onboard"
+                      - "SendAsync"
+                      - "Claim"
+                      - "CollaborativeRedeem"
+                      - "UnilateralExit"
+                      - "Balance"
+                      - "Stats"
+                  amount:
+                    type: number
+                    minimum: 0
+                  to:
+                    type: string
