--- simulation/schema.yaml
@@ -1,74 +1,0 @@
-$schema: "http://json-schema.org/draft-07/schema#"
-type: object
-required:
-  - version
-  - server
-  - clients
-  - rounds
-properties:
-  version:
-    type: string
-  server:
-    type: object
-    required:
-      - network
-      - round_interval
-    properties:
-      network:
-        type: string
-        enum: ["regtest", "testnet"]
-      round_interval:
-        type: integer
-        minimum: 1
-  clients:
-    type: array
-    items:
-      type: object
-      required:
-        - id
-      properties:
-        id:
-          type: string
-        name:
-          type: string
-        initial_funding:
-          type: number
-          minimum: 0
-  rounds:
-    type: array
-    items:
-      type: object
-      required:
-        - number
-        - actions
-      properties:
-        number:
-          type: integer
-          minimum: 1
-        actions:
-          type: object
-          patternProperties:
-            "^client_\\d+$":
-              type: array
-              items:
-                type: object
-                required:
-                  - type
-                properties:
-                  type:
-                    type: string
-                    enum:
-                      - "Onboard"
-                      - "SendAsync"
-                      - "SendOffchain"
-                      - "SendOnchain"
-                      - "Claim"
-                      - "CollaborativeRedeem"
-                      - "UnilateralExit"
-                      - "Balance"
-                      - "Stats"
-                  amount:
-                    type: number
-                    minimum: 0
-                  to:
-                    type: string
